Fix paths when user enters them.
