+++ render_job_to_rss_test.py
@@ -9,21 +9,6 @@ from proto import state_pb2
 class TestRenderJobToRss(unittest.TestCase):
   """Tests for the ShotNameBuilder class."""
 
-  def test_normalize_drive_letter(self):
-    """Test that the drive letter is normalized correctly."""
-    self.assertEqual(render_job_to_rss.normalize_drive_letter(
-        "c:/Users/peter/Documents/repositories/RenderRob/render_job_to_rss.py"),
-        "C:/Users/peter/Documents/repositories/RenderRob/render_job_to_rss.py")
-    self.assertEqual(render_job_to_rss.normalize_drive_letter(
-        "d:/test/file.txt"),
-        "D:/test/file.txt")
-    self.assertEqual(render_job_to_rss.normalize_drive_letter(
-        "e:\\test\\file.txt"),
-        "E:/test/file.txt")
-    self.assertEqual(render_job_to_rss.normalize_drive_letter(
-        "\\\\server\\share\\file.txt"),
-        "//server/share/file.txt")
-
   def test_render_job_to_render_settings_setter(self) -> None:
     """Test that the shot name is built correctly."""
     self.maxDiff = None  # pylint: disable=invalid-name
