+++ renderrob.py
@@ -6,13 +6,14 @@ import sys
 
 from PySide6.QtCore import QCoreApplication, QProcess, Qt
 from PySide6.QtGui import QAction, QColor, QIcon, QTextCharFormat, QTextCursor
-from PySide6.QtWidgets import QApplication, QFileDialog, QMessageBox, QWidget, QStackedLayout
+from PySide6.QtWidgets import QApplication, QFileDialog, QMessageBox, QWidget, QStackedLayout, QTableWidgetItem
 
 import settings_window
 import shot_name_builder
 from proto import cache_pb2, state_pb2
 from render_job_to_rss import render_job_to_render_settings_setter
 from state_saver import STATESAVER
+from tomlkit import table
 from utils import print_utils, table_utils, ui_utils
 
 MAX_NUMBER_OF_RECENT_FILES = 5
@@ -228,7 +229,15 @@ class MainWindow(QWidget):
     self.window.actionSettings.triggered.connect(self.open_settings_window)
     self.window.actionNew.triggered.connect(self.new_file)
     self.window.actionQuit.triggered.connect(self.quit)
+    self.table.itemChanged.connect(self.table_item_changed)
     #  #20 Add open blender button
+
+  def table_item_changed(self, item: QTableWidgetItem) -> None:
+    """Handle table item changes."""
+    # del item
+    self.is_saved = False
+    print(item)
+    table_utils.fix_active_row_path(item)
 
   def _handle_output(self):
     """Output the subprocess output to the textbrowser widget."""
